pyproject.toml
[project]
name = "cinemalibrarybot"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "attrs==25.4.0",
    "beautifulsoup4==4.14.3",
    "cachetools==6.2.1",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "dotenv==0.9.9",
    "frozenlist==1.8.0",
    "gemini-srt-translator==2.1.5",
    "google-auth==2.43.0",
    "google-genai==1.39.1",
    "greenlet==3.2.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "json-repair==0.51.0",
    "kurigram>=2.2.15",
    "multidict==6.7.0",
    "opensubtitlescom==0.1.5",
    "packaging==25.0",
    "prettytable==3.17.0",
    "propcache==0.4.1",
    "pyaes==1.6.1",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.3",
    "pydantic-core==2.33.1",
    "pysocks==1.7.1",
    "python-dotenv==1.1.0",
    "requests==2.32.5",
    "rsa==4.9.1",
    "sniffio==1.3.1",
    "soupsieve==2.8",
    "sqlalchemy==2.0.40",
    "sqlmodel==0.0.24",
    "srt==3.5.3",
    "tenacity==9.1.2",
    "tgcrypto-pyrofork==1.2.8",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "urllib3==2.5.0",
    "wcwidth==0.2.14",
    "websockets==15.0.1",
    "yarl==1.22.0",
]
